Move image helper classes to their own file
